Replace Values for some keys given a map
This functionallity adds to JoltUtils the capability to change the value
of one or more nodes given a map where the key match with a node key, and the
value is the new content for that node.key.

Two methods added:

JoltUtils.replaceRecursive -> Which receive a Jackson Object a Key and a value to replace
JoltUtils.replaceValues    -> Which receive a Jackson Object and a Map with several Keys and Values to replace.

Example:

Given:

{
        "L1_A": {
                "L2_A": {
                        "L3_A": "Good",
                        "L3_B": "RemoveThis"
                },
                "L2_B": "l2_b"
        },
        "L1_B": "l1_b",

        "L3_B": "RemoveThis"
}

When:

// java code

Map<String, Object> params = new HashMap<>();
params.put("L3_A", String.valueOf(j));
params.put("L3_B", String.valueOf(j));
// data is a Jackson Object
JoltUtils.replaceValues(data, params);

Then:

{
        "L1_A": {
                "L2_A": {
                        "L3_A": "Replaced1",
                        "L3_B": "Replaced2"
                },
                "L2_B": "l2_b"
        },
        "L1_B": "l1_b",

        "L3_B": "Replaced2"
}