Replace Values refactor:
Now we can provide the full path where the replacement will be done, and a map to match the values to be replaced.

Given the next example:

{
        "input" : {
            "L1_A" : {
                "L2_A" : {
                    "L3_A" : "Good",
                    "L3_B" : "ReplaceThis"
                },
                "L2_B" : "l2_b"
            },
            "L1_B" : "l1_b",

            "L3_B" : "This not should be replaced"
        },

        "mappingPaths" : {
            "L3_B" : ["L1_A.L2_A.L3_B"]
        },

        "valuesToReplace" : {
            "L3_B" : {
                "ReplaceThis" : "This has been replaced."
            }

        },

        "expected" : {
            "L1_A" : {
                "L2_A" : {
                    "L3_A" : "Good",
                    "L3_B" : "This has been replaced."
                },
                "L2_B" : "l2_b"
            },
            "L1_B" : "l1_b",

            "L3_B" : "This not should be replaced"
        }
    }

The 'input' is the original json, the 'mappingPaths' is a map where each key contains a list of fullpaths (a fullpath is the human reading representation of a node from json object)
the 'values' to replace is a map that contains a map where each key contains a map that represents the { "oldValue" : "newValue" } elements if the old value matches in the key where
then it will be replaced with the new value, if not it will be ignored.