Fix TOC links in README
GitHub automatically generates anchors for headings in markdown documents.
The links in the README table of contents do not match the pattern used by
GitHub so that are not useful when navigating README.md on GitHub.

This commit replaces the links with the correct anchors so that they will
work as intended.